# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  model.CreatePartnerSwagger:
    properties:
      image:
        example: ""
        type: string
      title:
        example: ""
        type: string
    type: object
...
  model.CreateProjectSwagger:
    properties:
      age_category:
        example: ""
        type: string
      category:
        example: ""
        type: string
      description:
        example: ""
        type: string
      image:
        example: ""
        type: string
      link:
        example: ""
        type: string
      period:
        example:
        - ""
        items:
          type: string
        type: array
      title:
        example: ""
        type: string
    type: object
...
  model.Project:
    properties:
      age_category:
        example: ""
        type: string
      category:
        example: ""
        type: string
      created:
        example: ""
        type: string
      description:
        example: ""
        type: string
      id:
        example: ""
        type: string
      image:
        example: ""
        type: string
      is_active:
        example: true
        type: boolean
      last_modified:
        example: ""
        type: string
      link:
        example: ""
        type: string
      period:
        example:
        - ""
        items:
          type: string
        type: array
      title:
        example: ""
        type: string
    required:
    - age_category
    - category
    - description
    - image
    - link
    - period
    - title
    type: object
  model.Response:
    properties:
      code:
        type: integer
      messsage:
        type: string
    type: object
...
  model.UpdatePartnerSwagger:
    properties:
      created:
        example: ""
        type: string
      id:
        example: ""
        type: string
      image:
        example: ""
        type: string
      title:
        example: ""
        type: string
    type: object
...
host: walking-school.site
info:
  contact:
    email: etverya12@gmail.com
    name: Yehor Tverytinov
  title: Walking-School backend API
  version: "1.0"
paths:
...
  /partner:
    get:
      consumes:
      - application/json
      parameters:
      - description: limit
        in: query
        name: limit
        type: string
      - description: offset
        in: query
        name: offset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get all partners.
      tags:
      - partner
    post:
      consumes:
      - application/json
      parameters:
      - description: Partner
        in: body
        name: Partner
        required: true
        schema:
          $ref: '#/definitions/model.CreatePartnerSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Create partner .
      tags:
      - partner
    put:
      consumes:
      - application/json
      parameters:
      - description: Partner
        in: body
        name: Partner
        required: true
        schema:
          $ref: '#/definitions/model.UpdatePartnerSwagger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Update partner by id.
      tags:
      - partner
  /partner/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Delete partner by id.
      tags:
      - partner
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get partner by id.
      tags:
      - partner
  /project:
    get:
      consumes:
      - application/json
      parameters:
      - description: limit
        in: query
        name: limit
        type: string
      - description: offset
        in: query
        name: offset
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "204":
          description: No Content
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get all projects.
      tags:
      - project
    post:
      consumes:
      - application/json
      parameters:
      - description: Project
        in: body
        name: Project
        required: true
        schema:
          $ref: '#/definitions/model.CreateProjectSwagger'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Create project.
      tags:
      - project
    put:
      consumes:
      - application/json
      parameters:
      - description: Project
        in: body
        name: Project
        required: true
        schema:
          $ref: '#/definitions/model.Project'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Update project by id.
      tags:
      - project
...
  /project/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Delete project by id.
      tags:
      - project
    get:
      consumes:
      - application/json
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.Response'
        "408":
          description: Request Timeout
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: Get project by id.
      tags:
      - project
...
swagger: "2.0"